pyproject.toml
[tool.poetry]
name = "crawl-news"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Scrapy = "^2.11.2"
requests = "^2.32.3"
lxml-html-clean = "^0.1.1"
beautifulsoup4 = "^4.12.3"
scrapy-splash = "^0.9.0"
feedparser = "^6.0.11"
pika = "^1.3.2"
aio-pika = "^9.4.2"
aiohttp = "^3.10.2"
google-generativeai = "^0.7.2"
python-dotenv = "^1.0.1"
markdownify = "^0.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
